requirements.txt
Flask
psycopg2-binary
Flask-Cors
python-dotenv
